aula 5.11

+++ src/main/kotlin/br/com/alura/alugames/modelo/InfoJogoJson.kt
@@ -1,8 +1,10 @@
 package br.com.alura.alugames.modelo
 
+import java.math.BigDecimal
+
 data class InfoJogoJson(
     val titulo: String,
     val capa: String,
-    val preco: Double,
+    val preco: BigDecimal,
     val descricao: String
 )
